% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calc_academic_year}
\alias{calc_academic_year}
\title{Calculate School or Academic Year from date}
\usage{
calc_academic_year(x, format = "long", date_parser = lubridate::ymd)
}
\arguments{
\item{x}{the date to be parsed}

\item{format}{one of `long`, `short`, `first_year`, or `second_year`,
which are rendered as `SY 2015-2016`, `15-16`, `2015`, or `2016`, respectively,
for the 2015-2016 school year.}

\item{date_parser}{a parser (such as as \code{lubrdiate}'s \code{\link[lubridate]{ymd}} function)}
}
\value{
a character or integer vector
}
\description{
Calculate School or Academic Year from date
}
\examples{
calc_academic_year('2014-09-15')
}

