% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/silo_connect.R
\name{connect_to_db}
\alias{connect_to_db}
\title{Create connection to Silo warehouse and return object to global environment}
\usage{
connect_to_db(db_name)
}
\arguments{
\item{db_name}{database name which will be connected to.}
}
\value{
a
}
\description{
This makes a connection to Silo via `RSQLServer`.  This call requires
that system environment variable for Silo are set (typically with a an `.Renviron`
file in the users home directory that is sourced when R is started).  This
function is called for it's side effect: createing a connection object for
the named database that can be used by `dplyr` calls in other functions in
this package.
}
\examples{

connect_to_db('Alumni_mirror')

}

