% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alumni_mirror.R
\name{get_alumni}
\alias{get_alumni}
\title{Collect table from Alumni mirror}
\usage{
get_alumni(table_name)
}
\arguments{
\item{table_name}{name of a table in illuminate mirror}
}
\value{
a dataframe
}
\description{
Collect table from Alumni mirror
}
