% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adb.R
\name{get_alumni_db}
\alias{get_alumni_db}
\title{Collect table from Alumni BQ mirror}
\usage{
get_alumni_db(table_name, collect = FALSE)
}
\arguments{
\item{table_name}{name of a table in illuminate mirror}

\item{collect}{if TRUE collect the table locally.}
}
\value{
a dataframe
}
\description{
Collect table from Alumni BQ mirror
}
