% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deanslist.R
\name{get_deanslist}
\alias{get_deanslist}
\title{Collect tables from DeansList on PARCC}
\usage{
get_deanslist(table_name = "suspensions", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from the DeansList Mirror on BigQuery}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect tables from DeansList on PARCC
}

