% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminate_mirror.R
\name{get_illuminate_mirror}
\alias{get_illuminate_mirror}
\title{Collect table from Illuminate_mirror}
\usage{
get_illuminate_mirror(table_name)
}
\arguments{
\item{table_name}{name of a table in illuminate mirror}
}
\value{
a dataframe
}
\description{
Collect table from Illuminate_mirror
}

