% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kickboard_mirror.R
\name{get_kb}
\alias{get_kb}
\title{Collect table from Kickboard db}
\usage{
get_kb(table_name)
}
\arguments{
\item{table_name}{name of a table in illuminate mirror}
}
\value{
a dataframe
}
\description{
Collect table from Kickboard db
}

