% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marts.R
\name{get_mart_students_attrition}
\alias{get_mart_students_attrition}
\title{Collect table from `mart_students_attendance` dataset on Big Query}
\usage{
get_mart_students_attrition(table_name = "transfers_by_month",
  collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from bq_mart_students_attendance dataset}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect table from `mart_students_attendance` dataset on Big Query
}
