% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwea.R
\name{get_nwea}
\alias{get_nwea}
\title{Collect table from NWEA}
\usage{
get_nwea(table_name)
}
\arguments{
\item{table_name}{name of a table from NWEA db}
}
\value{
a data.frame
}
\description{
Collect table from NWEA
}
