% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nwea_map_bq.R
\name{get_nwea_map}
\alias{get_nwea_map}
\title{Collect table from `nwea_map` dataset on Big Query}
\usage{
get_nwea_map(table_name = "cdf_combined_kipp_cps", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from PowerSchool dataset}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect table from `nwea_map` dataset on Big Query
}
