% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parcc.R
\name{get_parcc}
\alias{get_parcc}
\title{Collect table from PARCC}
\usage{
get_parcc(table_name = "parcc_student_results", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from PARCC dataset}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect table from PARCC
}
