% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_powerschool_bq.R
\name{get_powerschool}
\alias{get_powerschool}
\title{Collect table from PowerSchool on Big Query}
\usage{
get_powerschool(table_name = "students", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from PowerSchool dataset}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect table from PowerSchool on Big Query
}

