% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerschool_mirror.R
\name{get_ps}
\alias{get_ps}
\title{Collect table from PowerSchool_mirror}
\usage{
get_ps(table_name)
}
\arguments{
\item{table_name}{name of a table in alumni mirror}
}
\value{
a data.frame
}
\description{
Collect table from PowerSchool_mirror
}
