% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teachboost.R
\name{get_teachboost}
\alias{get_teachboost}
\title{Collect tables from TeachBoost}
\usage{
get_teachboost(table_name = "forms", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from the \code{teachboost} dataset on BigQuery}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect tables from TeachBoost
}

