% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tntp.R
\name{get_tntp}
\alias{get_tntp}
\title{Collect table from `tntp` dataset on Big Query}
\usage{
get_tntp(table_name = "items", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from TNTP dataset}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect table from `tntp` dataset on Big Query
}
