% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whetstone.R
\name{get_whetstone}
\alias{get_whetstone}
\title{Collect table from `whetstone` dataset on Big Query}
\usage{
get_whetstone(table_name = "observations", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table from Whetstone dataset}

\item{collect}{whether or not table should be pulled into local R session.
Default is `FALSE`.}
}
\value{
a data.frame
}
\description{
Collect table from `whetstone` dataset on Big Query
}
