% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminate_bq.R
\name{get_illuminate}
\alias{get_illuminate}
\title{Collect table from Alumni mirror}
\usage{
get_illuminate(table_name, schema = "dna_assessments", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table in illuminate}

\item{schema}{name of a table in illuminate schema that's
appended to the  dataset name (e.g., `public` for `illuminate_public`,
`dna` for `illuminate_dna`)}

\item{collect}{Whether to maintain data in BigQuery (`TRUE`) or to pull down
to your local environment (`FALSE`)}
}
\value{
a dataframe
}
\description{
Collect table from Alumni mirror
}
