% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/illuminate_bq.R
\name{get_illuminate}
\alias{get_illuminate}
\title{Collect table from Alumni mirror}
\usage{
get_illuminate(table_name, schema = "dna_assessments", collect = FALSE)
}
\arguments{
\item{table_name}{name of a table in illuminate mirror}
}
\value{
a dataframe
}
\description{
Collect table from Alumni mirror
}

